requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
Werkzeug==2.0.1
WTForms==2.3.3
